+++ linux/rollup/http/rollup-http-server/src/config.rs
@@ -15,17 +15,13 @@
 pub struct Config {
     pub http_address: String,
     pub http_port: u16,
-    pub dapp_http_address: String,
-    pub dapp_http_port: u16,
 }
 
 impl Config {
     pub fn new() -> Self {
         Self {
             http_address: String::from("127.0.0.1"),
-            http_port: 5001,
-            dapp_http_address: String::from("127.0.0.1"),
-            dapp_http_port: 5002,
+            http_port: 5004,
         }
     }
 }